# services/external/chat-gateway/api/server/api.yaml
openapi: 3.0.3
info:
  title: Chat Gateway API
  version: 1.0.0
servers:
  - url: /

tags:
  - name: auth
  - name: messages

paths:
  /api/v1/auth/register:
    post:
      operationId: AuthRegister
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        default:
          $ref: "#/components/responses/Error"

  /api/v1/auth/login:
    post:
      operationId: AuthLogin
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
          headers:
            Set-Cookie:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Error"

  /api/v1/auth/logout:
    post:
      operationId: AuthLogout
      tags: [auth]
      security:
        - cookieAuth: []
      responses:
        "204":
          description: No Content
        default:
          $ref: "#/components/responses/Error"

  /api/v1/messages:
    get:
      operationId: MessagesList
      tags: [messages]
      parameters:
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListMessagesResponse" }
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: MessagesCreate
      tags: [messages]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMessageRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        default:
          $ref: "#/components/responses/Error"

  /api/v1/messages/{id}:
    delete:
      operationId: MessagesDelete
      tags: [messages]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "204":
          description: No Content
        default:
          $ref: "#/components/responses/Error"

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid

  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }

  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [message, code]
      properties:
        message: { type: string }
        code: { type: string }

    RegisterRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 64 }
        password: { type: string, minLength: 8, maxLength: 256 }

    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 64 }
        password: { type: string, minLength: 8, maxLength: 256 }

    User:
      type: object
      additionalProperties: false
      required: [id, username, created_at]
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        created_at: { type: string, format: date-time }

    CreateMessageRequest:
      type: object
      additionalProperties: false
      required: [text]
      properties:
        text: { type: string, minLength: 1, maxLength: 2000 }

    Message:
      type: object
      additionalProperties: false
      required: [id, user_id, text, created_at]
      properties:
        id: { type: integer, format: int64 }
        user_id: { type: integer, format: int64 }
        text: { type: string }
        created_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }

    ListMessagesResponse:
      type: object
      additionalProperties: false
      required: [messages]
      properties:
        messages:
          type: array
          items: { $ref: "#/components/schemas/Message" }