# services/external/chat-gateway/internal/transport/http/generated/openapi.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package generated

import (
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v5"
	"github.com/oapi-codegen/runtime"
)

const (
	CookieAuthScopes = "cookieAuth.Scopes"
)

// CreateMessageRequest defines model for CreateMessageRequest.
type CreateMessageRequest struct {
	Text string `json:"text"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// ListMessagesResponse defines model for ListMessagesResponse.
type ListMessagesResponse struct {
	Messages []Message `json:"messages"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// Message defines model for Message.
type Message struct {
	CreatedAt time.Time  `json:"created_at"`
	DeletedAt *time.Time `json:"deleted_at"`
	Id        int64      `json:"id"`
	Text      string     `json:"text"`
	UserId    int64      `json:"user_id"`
}

// RegisterRequest defines model for RegisterRequest.
type RegisterRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// User defines model for User.
type User struct {
	CreatedAt time.Time `json:"created_at"`
	Id        int64     `json:"id"`
	Username  string    `json:"username"`
}

// Error defines model for Error.
type Error = ErrorResponse

// MessagesListParams defines parameters for MessagesList.
type MessagesListParams struct {
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// AuthLoginJSONRequestBody defines body for AuthLogin for application/json ContentType.
type AuthLoginJSONRequestBody = LoginRequest

// AuthRegisterJSONRequestBody defines body for AuthRegister for application/json ContentType.
type AuthRegisterJSONRequestBody = RegisterRequest

// MessagesCreateJSONRequestBody defines body for MessagesCreate for application/json ContentType.
type MessagesCreateJSONRequestBody = CreateMessageRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /api/v1/auth/login)
	AuthLogin(ctx *echo.Context) error

	// (POST /api/v1/auth/logout)
	AuthLogout(ctx *echo.Context) error

	// (POST /api/v1/auth/register)
	AuthRegister(ctx *echo.Context) error

	// (GET /api/v1/messages)
	MessagesList(ctx *echo.Context, params MessagesListParams) error

	// (POST /api/v1/messages)
	MessagesCreate(ctx *echo.Context) error

	// (DELETE /api/v1/messages/{id})
	MessagesDelete(ctx *echo.Context, id int) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// AuthLogin converts echo context to params.
func (w *ServerInterfaceWrapper) AuthLogin(ctx *echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.AuthLogin(ctx)
	return err
}

// AuthLogout converts echo context to params.
func (w *ServerInterfaceWrapper) AuthLogout(ctx *echo.Context) error {
	var err error

	ctx.Set(CookieAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.AuthLogout(ctx)
	return err
}

// AuthRegister converts echo context to params.
func (w *ServerInterfaceWrapper) AuthRegister(ctx *echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.AuthRegister(ctx)
	return err
}

// MessagesList converts echo context to params.
func (w *ServerInterfaceWrapper) MessagesList(ctx *echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params MessagesListParams
	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", ctx.QueryParams(), &params.Limit)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter limit: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.MessagesList(ctx, params)
	return err
}

// MessagesCreate converts echo context to params.
func (w *ServerInterfaceWrapper) MessagesCreate(ctx *echo.Context) error {
	var err error

	ctx.Set(CookieAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.MessagesCreate(ctx)
	return err
}

// MessagesDelete converts echo context to params.
func (w *ServerInterfaceWrapper) MessagesDelete(ctx *echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id int

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(CookieAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.MessagesDelete(ctx, id)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/api/v1/auth/login", wrapper.AuthLogin)
	router.POST(baseURL+"/api/v1/auth/logout", wrapper.AuthLogout)
	router.POST(baseURL+"/api/v1/auth/register", wrapper.AuthRegister)
	router.GET(baseURL+"/api/v1/messages", wrapper.MessagesList)
	router.POST(baseURL+"/api/v1/messages", wrapper.MessagesCreate)
	router.DELETE(baseURL+"/api/v1/messages/:id", wrapper.MessagesDelete)

}